+++ Tests/Wire/Resources/Listing/Listings/ListingsClientWireTests.swift
@@ -458,6 +458,12 @@ import Chrt
             sortOrder: .asc,
             page: 1,
             pageSize: 1,
+            filterStatus: [
+                .open
+            ],
+            filterType: [
+                .dispatch
+            ],
             filterTaskGroupId: "filter_task_group_id",
             filterOrderId: "filter_order_id",
             filterCreatedByOrgId: "filter_created_by_org_id",
@@ -581,6 +587,12 @@ import Chrt
             sortOrder: .asc,
             page: 1,
             pageSize: 1,
+            filterStatus: [
+                .open
+            ],
+            filterType: [
+                .dispatch
+            ],
             filterTaskGroupId: "filter_task_group_id",
             filterOrderId: "filter_order_id",
             filterCreatedByOrgId: "filter_created_by_org_id",
@@ -704,6 +716,12 @@ import Chrt
             sortOrder: .asc,
             page: 1,
             pageSize: 1,
+            filterStatus: [
+                .open
+            ],
+            filterType: [
+                .dispatch
+            ],
             filterTaskGroupId: "filter_task_group_id",
             filterOrderId: "filter_order_id",
             filterCreatedByOrgId: "filter_created_by_org_id",